+++ lib/createStore.d.ts
@@ -1,9 +1,9 @@
-export declare const retrievePersistedStore: (storageEngine: any) => any;
-export declare const persistStore: (storageEngine: any, store: any) => any;
+export declare const readStore: (storageEngine: any) => any;
+export declare const writeStore: (storageEngine: any, store: any) => any;
 export default function createStore(storageEngine?: any): {
-    set(key: any, value: any): any;
+    set(key: any, value: any): void;
     get(key: any): any;
-    deleteProperty(key: any): any;
+    deleteProperty(key: any): void;
     /**
      * Provide a subscribe method on the store that will notify the provided
      * callback with the target object that will be changed (set/del) along