+++ lib/memoryStorageStore.d.ts
@@ -1,7 +1,7 @@
 declare const store: {
-    set(key: any, value: any): any;
+    set(key: any, value: any): void;
     get(key: any): any;
-    deleteProperty(key: any): any;
+    deleteProperty(key: any): void;
     subscribe(fn: any): {
         dispose(): void;
     };